generic BookDataGetters cleanup
