+++ src/deal-with-background-songs.ts
@@ -9,7 +9,7 @@ import {
   getCurrentSectionTracks,
 } from "./audio-crossfader"; // Adjust path as needed
 import { CURRENT_BOOK } from "./consts"; // Adjust path as needed
-import { getBackgroundSongsForBook, BackgroundSongSection } from "./getBackgroundSongsForBook"; // Adjust path and ensure type export
+import { getBackgroundSongsForBook, BackgroundSongSection } from "./genericBookDataGetters/getBackgroundSongsForBook"; // Adjust path and ensure type export
 import { getCurrentLocation } from "@/helpers/paragraphsNavigation"; // Adjust path as needed
 
 let isProcessingBackgroundSongs = false; // Module-level flag to prevent re-entrancy
