+++ src/genericBookDataGetters/getCutScenesForBook.ts
@@ -1,4 +1,5 @@
 export type CutScene = { chapter: number; paragraph: number; file: string; text: string; delayInMs?: number };
+
 export const getCutScenesForBook = (): CutScene[] => {
   return [];
 };
